requirements.txt
azure-eventhub==5.11.4
azure-identity==1.15.0
python-dotenv==1.0.0 